Pulse check: no changes to code.
I am almost certain that the playGame method (i.e the central component
of my game engine) is now FUBAR because of fundamental changes made to
the makeMove function. Couple this with a less temporally bounded game
implementation (game is moved forward with clicks rather than console
logs) and I think it is entirely possible that makeMove is now the core
functionality of the engine. Currently conceptualizing a new approach.