Everything is complete and ready to deploy. Game plays and scoreboard appends with no glitches.
To eliminate burndown code, I created headline objects above my board where
the messages being console logged in my burndown code could be displayed.
In the scoreboard, I created span objects which would be updated in the
same fashion.

With these changes, the synthesis of my UI and game engine is complete.
The playGame function is kept in as legacy code, as attempts to remove it
and reformat create more headaches than time allows me to deal with.

Deploying and moving on to next problem: Ensuring that only those who
have signed in can play the game.