Cosmetic changes made, created banners which display your status in the login process.
Would like to make it so that the hidden blocks vanish, but that is a
stretch goal.

With this change, this feature is fully implemented. Now able to move on
to implementation of API. Given the severity of a headache it has been
for others in class, worried I will not be able to complete it for friday.
Plan on working for the rest of today and tomorrow.