First draft of code for my update API methods complete. Have yet to test. Have eyeballed two potential issues:
1.) in api.js line: 25. It is entirely possible that my URL definition
is a bad one and fails to properly point to the correct game. I imagine
This will require rounds of tweaking.

2.) a more minor issue. In the call on line 183 of the engine, I pass
an uppercase version of the mark through the data parser. The documentation
requires me to pass a lowercase. I do not imagine this is quite as serious
an issue, as I passed it a board full of numbers in create and got the
correct result. This said, it is still worth keeping in mind.