Problem was that game was never placed in store, causing my attempt to call store.game.id to fail. Updated appropriately. While this fix should have worked
I was faced with the same error. Concluded the problem was that I included two different ajax calls in the same block of code, and the async execution caused them to trip over each other and prevent either from firing. I thus need a way to get the create method to fire before the update method so the update method has a valid target.