Working on resubmission. Currently tackling problems surrounding change password.
Moved login form to footer, ensuring it only becomes visible after
the player logs in. Also updated message to show when the player has
input the wrong password.

This should solve the problems which surround my change password function.
Most of these problems stem from a hasty implementation, and this is a
better thought out version.